Fix a bug where commit.compile_and_cache_or_clear was never run.
It was never run because it returned immediately if "ready_changed?". But the hook is a after_commit hook, and so ready_changed? will never be true.
This seems to have gone unnoticed because after_commit hooks were converted to after_save hooks before. Now, we stopped that conversion, and we use the test_after_commit gem.