Temporarily removes entities from being fenced
This is required to ensure that we aren't forcing the translated copy
from using the same set of faced entities as the source copy.  For
example, in French, ampersands are rarely used.

In future, we should add support to distinguish between required vs
optional fences. For the time being, I've removed entities from being
fenced.

A comment is left there to mark that this should be done.